+++ android-sdk-ui/src/com/appboy/ui/inappmessage/views/AppboyInAppMessageModalView.java
@@ -80,7 +80,7 @@ public class AppboyInAppMessageModalView extends AppboyInAppMessageImmersiveBase
     scrollViewChild.setOnClickListener(new OnClickListener() {
       @Override
       public void onClick(View scrollView) {
-        AppboyLogger.d(TAG, String.format("Passing scrollView click event to message clickable view."));
+        AppboyLogger.d(TAG, "Passing scrollView click event to message clickable view.");
         getMessageClickableView().performClick();
       }
     });