+++ droidboy/src/com/appboy/sample/FeedCategoriesFragment.java
@@ -25,7 +25,7 @@ public class FeedCategoriesFragment extends DialogFragment {
      * implement this interface in order to receive event callbacks.
      * Each method passes the DialogFragment in case the host needs to query it. */
   public interface NoticeDialogListener {
-    public void onDialogPositiveClick(FeedCategoriesFragment dialog);
+    void onDialogPositiveClick(FeedCategoriesFragment dialog);
   }
 
   public FeedCategoriesFragment() {}