+++ wear-library/src/main/java/com/appboy/wear/enums/Month.java
@@ -15,5 +15,5 @@ public enum Month {
   SEPTEMBER,
   OCTOBER,
   NOVEMBER,
-  DECEMBER;
+  DECEMBER
 }