+++ wear-library/src/main/java/com/appboy/wear/enums/SocialNetwork.java
@@ -1,5 +1,5 @@
 package com.appboy.wear.enums;
 
 public enum SocialNetwork {
-  FACEBOOK, TWITTER, GOOGLE;
+  FACEBOOK, TWITTER, GOOGLE
 }